backend/requirements.txt
fastapi
uvicorn
networkx
google-generativeai
python-dotenv
PyGithub
pydantic
pandas